Move IndexedRepoRevs into internal/search/zoekt
