+++ internal/search/run/zoekt.go
@@ -3,7 +3,6 @@ package run
 import (
 	"context"
 	"fmt"
-	"strings"
 	"sync"
 	"time"
 	"unicode/utf8"
@@ -56,7 +55,7 @@ type IndexedSearchRequest struct {
 
 	// repos is the repository revisions that are indexed and will be
 	// searched.
-	repos *indexedRepoRevs
+	repos *zoektutil.IndexedRepoRevs
 
 	// since if non-nil will be used instead of time.Since. For tests
 	since func(time.Time) time.Duration
@@ -137,7 +136,7 @@ func NewIndexedSearchRequest(ctx context.Context, args *search.TextParameters, t
 	indexed, searcherRepos := zoektIndexedRepos(indexedSet, repos, filter)
 
 	tr.LogFields(
-		log.Int("indexed.size", len(indexed.repoRevs)),
+		log.Int("indexed.size", len(indexed.RepoRevs)),
 		log.Int("searcher_repos.size", len(searcherRepos)),
 	)
 
@@ -163,7 +162,7 @@ func (s *IndexedSearchRequest) Repos() map[string]*search.RepositoryRevisions {
 	if s.repos == nil {
 		return nil
 	}
-	return s.repos.repoRevs
+	return s.repos.RepoRevs
 }
 
 // Search streams 0 or more events to c.
@@ -188,11 +187,11 @@ func (s *IndexedSearchRequest) Search(ctx context.Context, c streaming.Sender) e
 // Timeouts are reported through the context, and as a special case errNoResultsInTimeout
 // is returned if no results are found in the given timeout (instead of the more common
 // case of finding partial or full results in the given timeout).
-func zoektSearch(ctx context.Context, args *search.TextParameters, repos *indexedRepoRevs, typ zoektutil.IndexedRequestType, since func(t time.Time) time.Duration, c streaming.Sender) error {
+func zoektSearch(ctx context.Context, args *search.TextParameters, repos *zoektutil.IndexedRepoRevs, typ zoektutil.IndexedRequestType, since func(t time.Time) time.Duration, c streaming.Sender) error {
 	if args == nil {
 		return nil
 	}
-	if len(repos.repoRevs) == 0 && args.Mode != search.ZoektGlobalSearch {
+	if len(repos.RepoRevs) == 0 && args.Mode != search.ZoektGlobalSearch {
 		return nil
 	}
 
@@ -211,10 +210,10 @@ func zoektSearch(ctx context.Context, args *search.TextParameters, repos *indexe
 	if args.Mode == search.ZoektGlobalSearch {
 		finalQuery = zoektquery.NewAnd(&zoektquery.Branch{Pattern: "HEAD", Exact: true}, queryExceptRepos)
 	} else {
-		finalQuery = zoektquery.NewAnd(&zoektquery.RepoBranches{Set: repos.repoBranches}, queryExceptRepos)
+		finalQuery = zoektquery.NewAnd(&zoektquery.RepoBranches{Set: repos.RepoBranches}, queryExceptRepos)
 	}
 
-	k := zoektutil.ResultCountFactor(len(repos.repoBranches), args.PatternInfo.FileMatchLimit, args.Mode == search.ZoektGlobalSearch)
+	k := zoektutil.ResultCountFactor(len(repos.RepoBranches), args.PatternInfo.FileMatchLimit, args.Mode == search.ZoektGlobalSearch)
 	searchOpts := zoektutil.SearchOpts(ctx, k, args.PatternInfo)
 
 	// Start event stream.
@@ -366,7 +365,7 @@ func zoektSearch(ctx context.Context, args *search.TextParameters, repos *indexe
 
 	mkStatusMap := func(mask search.RepoStatus) search.RepoStatusMap {
 		var statusMap search.RepoStatusMap
-		for _, r := range repos.repoRevs {
+		for _, r := range repos.RepoRevs {
 			statusMap.Update(r.Repo.ID, mask)
 		}
 		return statusMap
@@ -640,12 +639,12 @@ func queryToZoektQuery(query *search.TextPatternInfo, typ zoektutil.IndexedReque
 // zoektIndexedRepos splits the revs into two parts: (1) the repository
 // revisions in indexedSet (indexed) and (2) the repositories that are
 // unindexed.
-func zoektIndexedRepos(indexedSet map[string]*zoekt.Repository, revs []*search.RepositoryRevisions, filter func(*zoekt.Repository) bool) (indexed *indexedRepoRevs, unindexed []*search.RepositoryRevisions) {
+func zoektIndexedRepos(indexedSet map[string]*zoekt.Repository, revs []*search.RepositoryRevisions, filter func(*zoekt.Repository) bool) (indexed *zoektutil.IndexedRepoRevs, unindexed []*search.RepositoryRevisions) {
 	// PERF: If len(revs) is large, we expect to be doing an indexed
 	// search. So set indexed to the max size it can be to avoid growing.
-	indexed = &indexedRepoRevs{
-		repoRevs:     make(map[string]*search.RepositoryRevisions, len(revs)),
-		repoBranches: make(map[string][]string, len(revs)),
+	indexed = &zoektutil.IndexedRepoRevs{
+		RepoRevs:     make(map[string]*search.RepositoryRevisions, len(revs)),
+		RepoBranches: make(map[string][]string, len(revs)),
 	}
 	unindexed = make([]*search.RepositoryRevisions, 0)
 
@@ -665,119 +664,6 @@ func zoektIndexedRepos(indexedSet map[string]*zoekt.Repository, revs []*search.R
 	}
 
 	return indexed, unindexed
-}
-
-// indexedRepoRevs creates both the Sourcegraph and Zoekt representation of a
-// list of repository and refs to search.
-type indexedRepoRevs struct {
-	// repoRevs is the Sourcegraph representation of a the list of repoRevs
-	// repository and revisions to search.
-	repoRevs map[string]*search.RepositoryRevisions
-
-	// repoBranches will be used when we query zoekt. The order of branches
-	// must match that in a reporev such that we can map back results. IE this
-	// invariant is maintained:
-	//
-	//  repoBranches[reporev.Repo.Name][i] <-> reporev.Revs[i]
-	repoBranches map[string][]string
-}
-
-// headBranch is used as a singleton of the indexedRepoRevs.repoBranches to save
-// common-case allocations within indexedRepoRevs.Add.
-var headBranch = []string{"HEAD"}
-
-// Add will add reporev and repo to the list of repository and branches to
-// search if reporev's refs are a subset of repo's branches. It will return
-// the revision specifiers it can't add.
-func (rb *indexedRepoRevs) Add(reporev *search.RepositoryRevisions, repo *zoekt.Repository) []search.RevisionSpecifier {
-	// A repo should only appear once in revs. However, in case this
-	// invariant is broken we will treat later revs as if it isn't
-	// indexed.
-	if _, ok := rb.repoBranches[string(reporev.Repo.Name)]; ok {
-		return reporev.Revs
-	}
-
-	if !reporev.OnlyExplicit() {
-		// Contains a RefGlob or ExcludeRefGlob so we can't do indexed
-		// search on it.
-		//
-		// TODO we could only process the explicit revs and return the non
-		// explicit ones as unindexed.
-		return reporev.Revs
-	}
-
-	if len(reporev.Revs) == 1 && repo.Branches[0].Name == "HEAD" && (reporev.Revs[0].RevSpec == "" || reporev.Revs[0].RevSpec == "HEAD") {
-		rb.repoRevs[string(reporev.Repo.Name)] = reporev
-		rb.repoBranches[string(reporev.Repo.Name)] = headBranch
-		return nil
-	}
-
-	// Assume for large searches they will mostly involve indexed
-	// revisions, so just allocate that.
-	var unindexed []search.RevisionSpecifier
-	indexed := make([]search.RevisionSpecifier, 0, len(reporev.Revs))
-
-	branches := make([]string, 0, len(reporev.Revs))
-	for _, rev := range reporev.Revs {
-		if rev.RevSpec == "" || rev.RevSpec == "HEAD" {
-			// Zoekt convention that first branch is HEAD
-			branches = append(branches, repo.Branches[0].Name)
-			indexed = append(indexed, rev)
-			continue
-		}
-
-		found := false
-		for _, branch := range repo.Branches {
-			if branch.Name == rev.RevSpec {
-				branches = append(branches, branch.Name)
-				found = true
-				break
-			}
-			// Check if rev is an abbrev commit SHA
-			if len(rev.RevSpec) >= 4 && strings.HasPrefix(branch.Version, rev.RevSpec) {
-				branches = append(branches, branch.Name)
-				found = true
-				break
-			}
-		}
-
-		if found {
-			indexed = append(indexed, rev)
-		} else {
-			unindexed = append(unindexed, rev)
-		}
-	}
-
-	// We found indexed branches! Track them.
-	if len(indexed) > 0 {
-		rb.repoRevs[string(reporev.Repo.Name)] = reporev
-		rb.repoBranches[string(reporev.Repo.Name)] = branches
-	}
-
-	return unindexed
-}
-
-// GetRepoInputRev returns the repo and inputRev associated with file.
-func (rb *indexedRepoRevs) GetRepoInputRev(file *zoekt.FileMatch) (repo types.RepoName, inputRevs []string) {
-	repoRev := rb.repoRevs[file.Repository]
-
-	inputRevs = make([]string, 0, len(file.Branches))
-	for _, branch := range file.Branches {
-		for i, b := range rb.repoBranches[file.Repository] {
-			if branch == b {
-				// RevSpec is guaranteed to be explicit via zoektIndexedRepos
-				inputRevs = append(inputRevs, repoRev.Revs[i].RevSpec)
-			}
-		}
-	}
-
-	if len(inputRevs) == 0 {
-		// Did not find a match. This is unexpected, but we can fallback to
-		// file.Version to generate correct links.
-		inputRevs = append(inputRevs, file.Version)
-	}
-
-	return repoRev.Repo, inputRevs
 }
 
 // limitUnindexedRepos limits the number of repo@revs searched by the
